// web-application/src/types/preferences.ts
export interface UserPreferences {
  geminiApiKey: string 
  geminiModel: string 
  fishAudioApiKey: string 
  fishAudioVoiceId: string 
  showFurigana: boolean 
  audioSpeed: number 
  artworkDimming: number 
  streakCount: number 
  lastStreakDate: string 
  storiesReadCount: number 
  energyCount: number 
} 